service/service/requirements.txt
Django
djangorestframework
drf-spectacular
celery
redis
openai
requests
